+++ java-service-processor/src/test/resources/definition/TestIdMultiple.java
@@ -1,0 +1,11 @@
+package definition;
+
+import nbbrd.service.*;
+
+@ServiceDefinition(quantifier = Quantifier.MULTIPLE)
+interface TestIdMultiple {
+
+    @ServiceId
+    String getName();
+}
+